[level 2] Title: n^2 배열 자르기, Time: 10.45 ms, Memory: 92 MB -BaekjoonHub

+++ 프로그래머스/2/87390. n＾2 배열 자르기/n＾2 배열 자르기.java
@@ -2,7 +2,7 @@ import java.util.*;
 import java.io.*;
 class Solution {
     public int[] solution(int n, long left, long right) {
-        int size = (int)right - (int)left + 1;
+        int size = (int)(right - left + 1);
         int[] answer = new int[size];
     
         int index = 0;
